Adds simple auth middleware to all routes
