+++ compose.yml
@@ -20,6 +20,7 @@ services:
       - AUTH_PROVIDER_ACCEPTED_CLIENT_ID=1234
       - AUTH_PROVIDER_REDIRECT_URI=http://localhost:5173/oauth/cb
       - AUTH_PROVIDER_CODE_ENCRYPTION_SECRET=qM1Z5km9PF63E6uehLgJKztxpnhPE1Ag15Df1DwJrpY
+      - AUTH_INTERNAL_KEY=9876
       - DB_CONNECTION_STRING=postgresql://vivas:nosqueremos@db:5432/observatorio-femicidios
       - SERVER_CANNONICAL_ORIGIN=http://localhost:8080
